[DW-132] Adding missing beans
